change to double

+++ project_1.cpp
@@ -191,11 +191,11 @@ double TestSize(int n, int tests)
 
 double getMFlops(int test_size, double time_elapsed)
 {
-    long d = (double) test_size;
-    long num_multiplication =  d * d * d;
-    long num_additions = test_size * test_size * (test_size - 1);
+    double d = (double) test_size;
+    double num_multiplication =  d * d * d;
+    double num_additions = test_size * test_size * (test_size - 1);
     //    printf("#### %ld \t %ld ###\n", num_additions, num_multiplication);
-    long num_operations = num_additions+num_multiplication;
+    double num_operations = num_additions+num_multiplication;
     //    printf("#### %ld \t %f ###\n", num_operations,time_elapsed);
     double MFlops = num_operations / (time_elapsed * 1e6); // source used https://course.ccs.neu.edu/cs3650/ssl/TEXT-CD/Content/COD3e/InMoreDepth/IMD4-MFLOPS-as-a-Performance-Metric.pdf
